// resources/sap/cux/home/interface/AppsInterface.ts
export interface ICustomVisualization {
	appId?: string;
	oldAppId?: string;
	url?: string;
	leanURL?: string;
	title?: string;
	subtitle?: string;
	BGColor?: string | IColor;
	isFav?: boolean;
	isSection?: boolean;
	icon?: string;
	indicatorDataSource?: string;
	isCount?: boolean;
	contentProviderId?: string;
	isSmartBusinessTile?: boolean;
	visualization?: IVisualization;
	persConfig?: IPersConfig;
	setActive?: (val: boolean) => void;
	vizId?: string;
	status?: string;
	menuItems?: MenuItem[];
	addedInFavorites?: boolean;
	pageId?: string;
	refresh?: () => void;
	getDisplayFormat?: () => string;
	setLayoutData?: (data: GridContainerItemLayoutData) => void;
	bindProperty?: (propertyName: string, path: string) => void;
	fioriId?: string;
	vizInstance?: ICustomVizInstance;
}

export interface IPersConfig {
	pageId?: string;
	sectionIndex?: number;
	isDefaultSection?: boolean;
	visualizationIndex?: number;
	sectionId?: string;
	sectionTitle?: string;
	isPresetSection?: boolean;
	duplicateApps?: ICustomVisualization[];
}

export interface IAppInbounds {
	semanticObject: string;
	action: string;
	signature?: {
		parameters?: {
			"sap-fiori-id"?: {
				defaultValue: {
					value?: string;
				};
			};
		};
	};
	resolutionResult?: {
		applicationDependencies?: {
			name?: string;
			manifest?: ICardManifest;
		};
		ui5ComponentName?: string;
	};
}

export interface IVisualization {
	id?: string;
	icon?: string;
	info?: string;
	numberUnit?: string;
	vizConfig?: IVisualizationConfig;
	targetURL: string;
	title?: string;
	subtitle?: string;
	contentProviderId?: string;
	vizId: string;
	vizType: string;
	sectionId?: string;
	appIds?: string[];
	ignoreDuplicateApps?: boolean;
	_instantiationData?: {
		chip?: {
			configuration?: object;
			bags?: {
				sb_tileProperties?: {
					texts?: {
						title?: string;
						description?: string;
					};
				};
			};
		};
	};
	target?: {
		semanticObject?: string;
		action?: string;
	};
	indicatorDataSource?: {
		path: string;
	};
	isBookmark?: boolean;
	orgAppId?: string;
	displayFormatHint?: string;
	supportedDisplayFormats?: string;
	dataSource?: object;
}

export interface ISection {
	id?: string;
	index?: number;
	title?: string;
	isSection?: boolean;
	badge?: string;
	BGColor?: IColor | string;
	icon?: string;
	isPresetSection?: boolean;
	apps?: ICustomVisualization[];
	default?: boolean;
	preset?: boolean;
	visualizations?: IVisualization[];
	sectionIndex?: number;
	sectionProperties?: object;
}

export interface ICustomVizInstance extends Control {
	appId?: object;
	setActive: (active: boolean) => void;
	setClickable: (clickable: boolean) => void;
	getContent: () => {
		getComponentInstance: () => {
			getRootControl: () => Control;
		};
	};
}

export interface IVisualizationConfig {
	"sap.app": IVisualization;
	"sap.flp": IVisualization;
	"sap.ui": {
		icons: {
			icon: string;
		};
	};
}